+++ odoo/addons/ofh_sale_order_sap/models/__init__.py
@@ -8,3 +8,4 @@ from . import ofh_sale_order_line
 from . import ofh_payment_sap
 from . import ofh_payment
 from . import ofh_payment_request
+from . import common