+++ odoo/addons/ofh_sale_order_sap/wizards/__init__.py
@@ -1,0 +1,1 @@
+from . import ofh_sale_order_sap_import